pyproject.toml
[project]
name = "branch-flow"
version = "0.1.0"
dependencies = [
    "typer",
    "rich",
    "pyyaml"
]

[project.scripts]
bf = "branchflow.cli:app"

[project.optional-dependencies]
dev = [
    "ruff>=0.14.10",
    "mypy>=1.8.0",
    "pytest>=8.0"
]